refactor(slogger): break the Slogger/LogManager import cycle
`Slogger.ts` imported `LogManager` to resolve a handler config's
`type` / `formatter` strings, while `LogManager.ts` imported `Slogger`
as a value to construct its cached loggers — a two-way value-import
edge between the package's two central modules.

The handler/formatter registry now lives in a new leaf module,
`Registry.ts`, which both sides import instead:

    before:  Slogger.ts  <-->  LogManager.ts
    after:   Slogger.ts   -->  Registry.ts  <--  LogManager.ts

`Registry` owns the two maps and the `addHandler` / `addFormatter` /
`createFormatter` / `getFormatter` / `getFormatterNames` /
`getHandlerTypes` / `createHandler` logic, verbatim, including the
built-in registrations that used to run in `Manager`'s constructor.
`Manager` keeps the logger cache (`_loggers`, `createSlogger`,
`getLogger`, `__sameConfig`) — the half that genuinely needs `Slogger`
as a value — and forwards the registry half to the shared instance.
Its only back-edge to `Slogger.ts` is `import type { HandlerConfig }`,
which is erased at compile time and cannot create a runtime cycle.

Shared-singleton guarantee: the registry is a module-scope binding
(`export const registry: Registry = new Registry()`), so every importer
in the package shares one instance and one pair of maps. Delegation, not
inheritance, is what makes this hold — a `Manager extends Registry` would
have given `LogManager` its own maps and silently split lookup, so a
formatter added via `LogManager.addFormatter` would have been invisible
to a `Slogger` resolving `formatter: 'name'`. The new `Registry.test.ts`
pins this from both directions: it writes through one import path and
reads through another, and drives the end-to-end path where a handler
type registered via `LogManager` is resolved by a plain `new Slogger()`.

No public API change: `LogManager` is still exported from `LogManager.ts`
with an identical method surface, `mod.ts` is untouched, and the removed
`_handlers` / `_formatters` fields were `protected` on a class that was
never exported.

Why now — this is hardening, not a live bug fix. Circular ESM imports are
legal and work natively. But one top-level `await` anywhere in a
consumer's module graph makes esbuild (wrangler) and Rollup (Vite) lower
every module initializer in that graph to an async function, and a
two-way edge then deadlocks: `await init_Slogger()` awaits
`await init_LogManager()` which awaits `init_Slogger()`. slogger hung
forever on import in wrangler and Vite consumer builds until
`@tundralibs/compat` 1.1.2 removed the offending TLA. That hang is gone
today; removing the cycle makes slogger robust regardless of what a
consumer's graph contains, rather than dependent on compat staying
TLA-free.
